fix few more unit tests.
